+++ GladysCommon/Sources/GladysCommon/ArchivedItem.swift
@@ -81,8 +81,8 @@ public final class ArchivedItem: Codable {
 
     // Transient
     public struct Flags: OptionSet {
-        public let rawValue: UInt8
-        public init(rawValue: UInt8) {
+        public let rawValue: Int
+        public init(rawValue: Int) {
             self.rawValue = rawValue
         }
